use i32 instead of i64 as order type

+++ src/cl.rs
@@ -16,12 +16,12 @@ pub trait Cl<T> {
     ///
     /// println!("Cl_{}({}) = {}", 2, 1.0, 1.0.cl(2));
     /// ```
-    fn cl(&self, n: i64) -> T;
+    fn cl(&self, n: i32) -> T;
 }
 
 
 impl Cl<f64> for f64 {
-    fn cl(&self, n: i64) -> f64 {
+    fn cl(&self, n: i32) -> f64 {
         match n {
             1 => cl1::cl1(*self),
             2 => cl2::cl2(*self),
